Move game UI components to game folder
