+++ src/ui/game/shop_panel.cpp
