+++ src/ui/game/shop_panel.h
